Clean Up Documentation for Filter, VFS, and Group
* All low level wrapped functions have been made protected
* Clean up and add missing documentation
* Move tests for classes no longer in libtiledb.pyx into new files
